To avoid the problem of accepting the ratio of the player’s image when
adding an image in "add player", it is necessary to create a condition

that said that if the quotient of the width and height of the image are

not equal to 1 then we put an exception that warn that the image does

not have the correct ratio.

Moreover, in the controller you must create a dialog message that works

only if the user clicks on save in the Jfilechooser.

Alfred Yael